Updated victory/loss conditions
for all conditions other than damage.

+++ Assets/Scripts/PlayerStats.cs
@@ -8,6 +8,6 @@ public class PlayerStats : ScriptableObject
 {
     public float MaxHealth = 10;
     public float CurrentHealth = 10;
-    public float CoinsCollected = 0;
-    public float CoinsLeft = 10;
+    public int CoinsCollected = 0;
+    public int CoinsLeft = 10;
 }
